Redid build dependency check system
The following scenario would cause issues in the previous system:

Have a package marked as a dependency (i.e. 'makedeb'), and have another 
package provide for that (i.e. 'makedeb-beta').

In the old version, makedeb would try to install 'makedeb', even if 
'makedeb-beta' was already installed.

The new version would do nothing when 'makedeb-beta' is installed, as 
'makedeb-beta' provides for the 'makedeb' package, thus resolving the 
dependency.