Fixed error when using the '-i' option; Did some code cleanup
Previously, convert_version() would fail if 'epoch' was not set due to 
the 'set -e' flag at the beginning of the script (combined with the '&&' 
operator which would exit with 1 if the epoch check failed). It's now 
placed inside of an if statement, which will just ignore any bad exit 
codes.