Fixed issue when extra control fields are specified
Previously, makedeb would add the "Adding field '{field}' to control 
file" text to the built control file, which would cause the resulting 
packages to be uninstallable. This was due to how generate_control()'s 
output was handled, which redirected everything from it into the control 
file, which would, as seen, include said text.

Now, the output file is simply passed as a parameter to 
'generate_control()', which allows for more control for which strings 
end up inside the output file.