Fixed issue with duplicated warnings in argument check
Argument checks were previously performed at the beginning of the 
script, as well as after entering the fakeroot environment, which caused 
things like warnings in the arg check to be printed twice.

Warnings and errors are now hidden in the arg check when inside the 
fakeroot, thus alleviating the issue.