Fixed .SRCINFO fixer command used for AUR deployments
This issue is a bit more complicated than others, but the gist is as 
follows (hopefully :P):

the 'cat' command doesn't have to complete before its output is sent 
through the pipes. Likewise, the 'tee' command starts overwriting the 
SRCINFO file before 'cat' is complete, thus doing the following:

- tee overwrites the file with nothing since it's empty (haven't 
checked, but that's my hypothesis) *before* cat is done reading the file

- cat then doesn't get all the previous data in a file, since it has 
been previously been overwritten by tee with nothing.

That's a lesson learned: don't read and write to the same file in a Bash 
pipe.