Fixed SSH config folder issues in CI script
So apparently running 'chmod 400' on a directory recursively is gonna 
cause commands like 'chmod' to not be able to enter the directory when 
they need to, which is causing issues when chmod is changing perms on 
the directory first, and then can't enter the directory to change 
permissions on enclosed files.

Don't know if that's what's happening, but it's what I'm observing.