Fixed issue with installing build dependencies when conflicting packages are present
Previously, when installing a build dependency that conflicted with a package already on the user's system (i.e. attempting to install 'makedeb-makepkg-beta' when 'makedeb-makepkg-alpha' is installed), apt-get would fail due to the use of the 'satisfy' command. The 'install' command is now used, which will also provide for more compatibility with older releases of APT.