Fixed command used to install build dependencies
Looks like sudo doesn't pass the user's environment to the commands it runs by default, and requires the '-E' option to do so.

Also adding the 'tzdata' to the list of dependencies installed before
makedeb is ran until we can add the 'sudo -E' functionality as a feature
flag to makedeb.