Temporarily disabled unit test for '-r' functionality
We need to package up Bats and some helper libraries before we can use this test. Our main issue right now is finding the correct output of dpkg when multiple lines are present, and while we could implement a function to search each line and see if one matches manually, it's gonna be easier if we just wait to package some Bats libraries that already support it.