Fix refs not getting checked out when specified in PKGBUILDs
Before the shallow cloning implementation, we were checking out to the reference *after* cloning, but since we only clone one commit, we need to that during cloning now instead.

Likewise, we didn't set the variable to check for the specified reference until after cloning as well, which caused issues with finding references during cloning in the new implementation.